UI rewrite and additional group/resource functions
This update was a full rewrite of the tkinter UI to allow for multiple PIM types (roles, groups, and resources). Roles and resources have been fully built out to allow full for activation and monitoring. You can only monitor groups for now. I've also included checking for reauth reequipments on certain roles if that setting is enabled. It will automatically for you back to the browser to fully authenticate again then reattempts activation. Many other quality of life features as well.